# Optimize filter operations