--- a/src/NUnitFramework/framework/Internal/Filters/AndFilter.cs
+++ b/src/NUnitFramework/framework/Internal/Filters/AndFilter.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
 #nullable enable
-using System.Linq;
 using NUnit.Framework.Interfaces;
 
 namespace NUnit.Framework.Internal.Filters
@@ -32,9 +31,27 @@ namespace NUnit.Framework.Internal.Filters
         public override bool Pass( ITest test, bool negated )
         {
             if (negated)
-                return Filters.Any(f => f.Pass(test, negated));
+            {
+                foreach (var filter in Filters)
+                {
+                    if (filter.Pass(test, negated))
+                    {
+                        return true;
+                    }
+                }
 
-            return Filters.All(f => f.Pass(test, negated));
+                return false;
+            }
+
+            foreach (var filter in Filters)
+            {
+                if (!filter.Pass(test, negated))
+                {
+                    return false;
+                }
+            }
+
+            return true;
         }
 
         /// <summary>
@@ -44,7 +61,15 @@ namespace NUnit.Framework.Internal.Filters
         /// <returns>True if all the component filters match, otherwise false</returns>
         public override bool Match( ITest test )
         {
-            return Filters.All(filter => filter.Match(test));
+            foreach (var filter in Filters)
+            {
+                if (!filter.Match(test))
+                {
+                    return false;
+                }
+            }
+
+            return true;
         }
 
         /// <summary>
@@ -54,9 +79,13 @@ namespace NUnit.Framework.Internal.Filters
         /// <returns>True if all the component filters explicit match, otherwise false</returns>
         public override bool IsExplicitMatch( ITest test )
         {
-            foreach( TestFilter filter in Filters )
-                if ( !filter.IsExplicitMatch( test ) )
+            foreach (var filter in Filters)
+            {
+                if (!filter.IsExplicitMatch(test))
+                {
                     return false;
+                }
+            }
 
             return true;
         }