--- a/src/NUnitFramework/framework/Internal/Filters/AndFilter.cs
+++ b/src/NUnitFramework/framework/Internal/Filters/AndFilter.cs
@@ -30,6 +30,8 @@ namespace NUnit.Framework.Internal.Filters
         /// <returns>True if all the component filters pass, otherwise false</returns>
         public override bool Pass( ITest test, bool negated )
         {
+            // Use foreach-loop against array instead of LINQ for best performance
+
             if (negated)
             {
                 foreach (var filter in Filters)
@@ -61,6 +63,7 @@ namespace NUnit.Framework.Internal.Filters
         /// <returns>True if all the component filters match, otherwise false</returns>
         public override bool Match( ITest test )
         {
+            // Use foreach-loop against array instead of LINQ for best performance
             foreach (var filter in Filters)
             {
                 if (!filter.Match(test))
@@ -79,6 +82,7 @@ namespace NUnit.Framework.Internal.Filters
         /// <returns>True if all the component filters explicit match, otherwise false</returns>
         public override bool IsExplicitMatch( ITest test )
         {
+            // Use foreach-loop against array instead of LINQ for best performance
             foreach (var filter in Filters)
             {
                 if (!filter.IsExplicitMatch(test))