--- a/src/NUnitFramework/framework/Internal/Filters/CategoryFilter.cs
+++ b/src/NUnitFramework/framework/Internal/Filters/CategoryFilter.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
 #nullable enable
-using System.Collections;
-using System.Linq;
 using NUnit.Framework.Interfaces;
 
 namespace NUnit.Framework.Internal.Filters
@@ -27,9 +25,18 @@ namespace NUnit.Framework.Internal.Filters
         /// <returns></returns>
         public override bool Match(ITest test)
         {
-            IList testCategories = test.Properties[PropertyNames.Category];
+            if (test.Properties.TryGet(PropertyNames.Category, out var testCategories))
+            {
+                for (var i = 0; i < testCategories.Count; ++i)
+                {
+                    if (Match((string) testCategories[i]))
+                    {
+                        return true;
+                    }
+                }
+            }
 
-            return testCategories.Cast<string>().Any(Match);
+            return false;
         }
 
         /// <summary>