--- a/src/NUnitFramework/framework/Internal/Filters/OrFilter.cs
+++ b/src/NUnitFramework/framework/Internal/Filters/OrFilter.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
 #nullable enable
-using System.Linq;
 using NUnit.Framework.Interfaces;
 
 namespace NUnit.Framework.Internal.Filters
@@ -66,7 +65,15 @@ namespace NUnit.Framework.Internal.Filters
         /// <returns>True if any of the component filters match, otherwise false</returns>
         public override bool Match( ITest test )
         {
-            return Filters.Any(filter => filter.Match(test));
+            foreach (var filter in Filters)
+            {
+                if (filter.Match(test))
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         /// <summary>
@@ -76,7 +83,15 @@ namespace NUnit.Framework.Internal.Filters
         /// <returns>True if any of the component filters explicit match, otherwise false</returns>
         public override bool IsExplicitMatch( ITest test )
         {
-            return Filters.Any(filter => filter.IsExplicitMatch(test));
+            foreach (var filter in Filters)
+            {
+                if (filter.IsExplicitMatch(test))
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         /// <summary>