--- a/src/NUnitFramework/framework/Internal/PreFilter.cs
+++ b/src/NUnitFramework/framework/Internal/PreFilter.cs
@@ -198,6 +198,7 @@ namespace NUnit.Framework.Internal
 
             private bool MatchSetUpFixture(Type type)
             {
+                // checking length instead of for example LINQ .Any(), which would need to box array into IEnumerable
                 return IsSubNamespace(type.Namespace) &&
                        type.GetCustomAttributes(typeof(SetUpFixtureAttribute), true).Length > 0;
             }