--- a/src/NUnitFramework/nunitlite.tests/CreateTestFilterTests.cs
+++ b/src/NUnitFramework/nunitlite.tests/CreateTestFilterTests.cs
@@ -31,7 +31,7 @@ namespace NUnitLite.Tests
             var filter = GetFilter("--test:My.First.Test", "--test:My.Second.Test", "--test:My.Third.Test");
             Assert.That(filter, Is.TypeOf<OrFilter>());
             var filters = ((OrFilter)filter).Filters;
-            Assert.That(filters.Count, Is.EqualTo(3));
+            Assert.That(filters.Length, Is.EqualTo(3));
 
             Assert.That(filters[0], Is.TypeOf<FullNameFilter>());
             Assert.That(((FullNameFilter)filters[0]).ExpectedValue, Is.EqualTo("My.First.Test"));
@@ -51,7 +51,7 @@ namespace NUnitLite.Tests
                 var filter = GetFilter("--testlist:" + tf.File.FullName);
                 Assert.That(filter, Is.TypeOf<OrFilter>());
                 var filters = ((OrFilter)filter).Filters;
-                Assert.That(filters.Count, Is.EqualTo(3));
+                Assert.That(filters.Length, Is.EqualTo(3));
 
                 Assert.That(filters[0], Is.TypeOf<FullNameFilter>());
                 Assert.That(((FullNameFilter)filters[0]).ExpectedValue, Is.EqualTo("My.First.Test"));
@@ -73,7 +73,7 @@ namespace NUnitLite.Tests
                 var filter = GetFilter("--testlist:" + tf.File.FullName, "--testlist:" + tf2.File.FullName );
                 Assert.That(filter, Is.TypeOf<OrFilter>());
                 var filters = ((OrFilter)filter).Filters;
-                Assert.That(filters.Count, Is.EqualTo(6));
+                Assert.That(filters.Length, Is.EqualTo(6));
 
                 Assert.That(filters[0], Is.TypeOf<FullNameFilter>());
                 Assert.That(((FullNameFilter)filters[0]).ExpectedValue, Is.EqualTo("My.First.Test"));
@@ -121,14 +121,14 @@ namespace NUnitLite.Tests
 
             Assert.That(filter, Is.TypeOf<AndFilter>());
             var filters = ((AndFilter)filter).Filters;
-            Assert.That(filters.Count, Is.EqualTo(2));
+            Assert.That(filters.Length, Is.EqualTo(2));
 
             Assert.That(filters[1], Is.TypeOf<CategoryFilter>());
             Assert.That(((CategoryFilter)filters[1]).ExpectedValue, Is.EqualTo("Urgent"));
 
             Assert.That(filters[0], Is.TypeOf<OrFilter>());
             filters = ((OrFilter)filters[0]).Filters;
-            Assert.That(filters.Count, Is.EqualTo(2));
+            Assert.That(filters.Length, Is.EqualTo(2));
 
             Assert.That(filters[0], Is.TypeOf<FullNameFilter>());
             Assert.That(((FullNameFilter)filters[0]).ExpectedValue, Is.EqualTo("My.First.Test"));