I was investigating how to make filtering and test case traversal faster as this is what I believe directly correlates with user experience and lag when things are being calculated and discovered. I tried to stay on filtering side for now, but I have other interesting branches as work in progress too 😉 

The `CountTestsWithCategoryFilter` numbers pretty nicely sum up the pitfalls of using LINQ in capturing context memory-wise (from 68MB to 12KB).

* change `CompositeFilter.Filters` to be an array for fast traversal - internal type so signature change should be OK
* replace LINQ usage with non-allocating plain loops
* add `TryGet` to `PropertyBag` which will not allocate on retrieval in contrast to indexer, name OK?
* detect empty filter from string without the need to parse the XML

I'm adding some comments to PR to clarify my suggestions/reasons behind the changes.

Following testing has been done against `Jint.Tests.Test262` test project as the source for discovery, which has the > 60k test cases to traverse.

Please also consider the delta of change instead of the original nano/milliseconds as I have quite beefy desktop - so halving the wait time on older hardware / CI pipelines can shine.

``` ini
BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.22621.1631/22H2/2022Update/SunValley2)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=7.0.203
  [Host] : .NET 6.0.16 (6.0.1623.17311), X64 RyuJIT AVX2

Job=InProcess  Toolchain=InProcessEmitToolchain  
```

## NUnit.Framework.FrameworkControllerBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |CountTestsEmptyFilter|960.0 μs|4.16 μs|47.38 KB|
| **New** |	| **736.3 μs (-23%)** | **9.67 μs** | **1 B (-100%)** |
| Old |CountTestsWithCategoryFilter|40,061.2 μs|764.54 μs|68642.27 KB|
| **New** |	| **15,705.2 μs (-61%)** | **191.34 μs** | **12974 B (-100%)** |
| Old |CountTestsWithOrFilterAll|72,789.7 μs|548.23 μs|35458.02 KB|
| **New** |	| **71,194.9 μs (-2%)** | **399.37 μs** | **36272121 B (0%)** |
| Old |LoadTests|337,160.2 μs|3,950.07 μs|208299.28 KB|
| **New** |	| **332,760.1 μs (-1%)** | **4,875.89 μs** | **213225540 B (0%)** |