@lahma 2023-04-23
Thank you for the review @OsirisTerje ! I tried to address the issues by adding comments why `for`/`foreach` was selected. Some of the remarks would probably require refactoring the class hierarchy / introducing new members so I didn't go there (yet at least).