@OsirisTerje 2023-04-23
@lahma 
Introducing new members - feel free to do that!  
Rearranging the hierarchy, well, we do have tests, don't we?  Might work.