@lahma 2023-04-23
@OsirisTerje I would be glad if this was good for now, maybe refactoring is for another time. Especially nice would be sealing the types that are not expected to be actually inherited outside of the framework - but probably a case of another PR.