@OsirisTerje 2023-04-23
If you like, you could create an issue, label it with is:refactoring, and just list the things you think should be done later.